feat: unify donation domain schema and services
