+++ apps/api/src/donations/donation-payout-state.enum.ts
@@ -1,0 +1,7 @@
+export enum DonationPayoutState {
+  Unassigned = "unassigned",
+  Scheduled = "scheduled",
+  Processing = "processing",
+  Paid = "paid",
+  Failed = "failed"
+}
