+++ apps/api/src/donations/donation-status.enum.ts
@@ -1,5 +1,7 @@
 export enum DonationStatus {
   Pending = "pending",
-  Paid = "paid",
-  Failed = "failed"
+  Processing = "processing",
+  Succeeded = "succeeded",
+  Failed = "failed",
+  Refunded = "refunded"
 }
