+++ apps/api/src/donations/models/donation.model.ts
@@ -1,9 +1,11 @@
 import { Field, ID, Int, ObjectType, registerEnumType } from "@nestjs/graphql";
 import { GraphQLISODateTime } from "@nestjs/graphql";
-import { DonationStatus } from "../donation.schema";
+import { DonationPayoutState } from "../donation-payout-state.enum";
+import { DonationStatus } from "../donation-status.enum";
 import { DonationStatusChangeModel } from "./donation-status-change.model";
 
 registerEnumType(DonationStatus, { name: "DonationStatus" });
+registerEnumType(DonationPayoutState, { name: "DonationPayoutState" });
 
 @ObjectType("Donation")
 export class DonationModel {
@@ -14,16 +16,37 @@ export class DonationModel {
   declare submissionId: string;
 
   @Field()
+  declare challengeId: string;
+
+  @Field()
+  declare creatorUserId: string;
+
+  @Field()
   declare donorUserId: string;
 
   @Field(() => Int)
   declare amountCents: number;
+
+  @Field(() => Int)
+  declare platformFeeCents: number;
+
+  @Field(() => Int)
+  declare creatorShareCents: number;
+
+  @Field(() => Int)
+  declare platformShareCents: number;
+
+  @Field(() => Int)
+  declare platformVatCents: number;
 
   @Field()
   declare currency: string;
 
   @Field(() => DonationStatus)
   declare status: DonationStatus;
+
+  @Field(() => DonationPayoutState)
+  declare payoutState: DonationPayoutState;
 
   @Field(() => [DonationStatusChangeModel])
   declare statusHistory: DonationStatusChangeModel[];
